Order upload and retry code
